+++ front/app/containers/Admin/projects/project/nativeSurvey/messages.ts
@@ -1,0 +1,139 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  survey2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.survey2',
+    defaultMessage: 'Survey',
+  },
+  surveyDescription2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.surveyDescription2',
+    defaultMessage: "Create and edit the questions in this projects's survey.",
+  },
+  downloadAllResults2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.downloadAllResults2',
+    defaultMessage: 'Download all survey results',
+  },
+  totalSurveyResponses2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.totalSurveyResponses2',
+    defaultMessage: 'Total {count} responses',
+  },
+  noSurveyResponses2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.noSurveyResponses2',
+    defaultMessage: 'No survey responses yet',
+  },
+  surveyResults2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.surveyResults2',
+    defaultMessage: 'Survey results',
+  },
+  downloadResults2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.downloadResults2',
+    defaultMessage: 'Download survey results',
+  },
+  informationText2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.informationText2',
+    defaultMessage:
+      'Responses to short and long answer questions are currently only available in the survey download',
+  },
+  required2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.required2',
+    defaultMessage: 'Required',
+  },
+  optional2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.optional2',
+    defaultMessage: 'Optional',
+  },
+  choiceCount2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.choiceCount2',
+    defaultMessage:
+      '{percentage}% ({choiceCount, plural, no {# choices} one {# choice} other {# choices}})',
+  },
+  // DeleteFormResultsNotice
+  openForResponses2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.openForResponses2',
+    defaultMessage: 'Open for responses',
+  },
+  deleteSurveyResults2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.deleteSurveyResults2',
+    defaultMessage: 'Delete survey results',
+  },
+  viewSurveyResults2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.viewSurveyResults2',
+    defaultMessage: 'View survey results ({count})',
+  },
+  editSurveyContent2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.editSurveyContent2',
+    defaultMessage: 'Edit survey content',
+  },
+  viewSurveyText2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.viewSurveyText2',
+    defaultMessage: 'View survey',
+  },
+  deleteResultsConfirmationQuestion2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.deleteSurveyResultsConfirmation2',
+    defaultMessage: 'Are you sure you want to delete all survey results?',
+  },
+  deleteResultsInfo2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.deleteResultsInfo2',
+    defaultMessage: 'This cannot be undone',
+  },
+  cancelDeleteButtonText2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.cancelDeleteButtonText2',
+    defaultMessage: 'Cancel',
+  },
+  confirmDeleteButtonText2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.confirmDeleteButtonText2',
+    defaultMessage: 'Yes, delete survey results',
+  },
+  disabledSurveyMessage2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.disabledSurveyMessage2',
+    defaultMessage:
+      "Survey content can't be edited as survey results have started coming in.",
+  },
+  deleteResultsLink2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.deleteResultsLink2',
+    defaultMessage: 'Delete the results if you need to make changes.',
+  },
+  viewSurvey2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.viewSurvey2',
+    defaultMessage: 'View survey',
+  },
+  addSurveyContent2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.addSurveyContent2',
+    defaultMessage: 'Add survey content',
+  },
+  successMessage2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.successMessage2',
+    defaultMessage: 'Survey successfully saved',
+  },
+  supportArticleLink2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.supportArticleLink2',
+    defaultMessage:
+      'https://support.citizenlab.co/en/articles/6673873-creating-an-in-platform-survey',
+  },
+  surveyEnd2: {
+    id: 'app.containers.AdminPage.ProjectEdit.formBuilder.surveyEnd2',
+    defaultMessage: 'Survey end',
+  },
+  questionLogicHelperText2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.questionLogicHelperText2',
+    defaultMessage:
+      'If no logic is added to an answer, the survey will follow its normal flow. Adding logic to a question will make it required by default. For more information, visit {supportPageLink}',
+  },
+  pagesLogicHelperText2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.pagesLogicHelperText2',
+    defaultMessage:
+      'If no logic is added, the survey will follow its normal flow. For more information, visit {supportPageLink}',
+  },
+  multiselect2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.multiselectText2',
+    defaultMessage: 'Multiple choice - choose many',
+  },
+  select2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.selectText2',
+    defaultMessage: 'Multiple choice - choose one',
+  },
+  linear_scale2: {
+    id: 'app.containers.AdminPage.ProjectEdit.survey.linear_scale2',
+    defaultMessage: 'Linear scale',
+  },
+});