+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2023-02-24
+
+### Changed
+
+- [CL-2903] Lock first and last name with FC verification
+- [CL-2721] Make it possible to move project out of folder
+- Release missing native survey translations
+
 ## 2023-02-22
 
 ### Changed